# src/app/docs/swagger.yaml
basePath: /
definitions:
  main.Beer:
    properties:
      abv:
        type: number
      ibu:
        type: number
      id:
        type: string
      name:
        type: string
      srm:
        type: number
      style:
        type: string
      type:
        type: string
      upc:
        type: number
      update:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: soberkoder@swagger.io
    name: API Support
  description: This is a sample serice for managing brewery
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sample API
  version: "1.0"
paths:
  /api/addbeer:
    post:
      consumes:
      - application/json
      description: add by json beer
      parameters:
      - description: Add Beer
        in: body
        name: beer
        required: true
        schema:
          $ref: '#/definitions/main.Beer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Beer'
      summary: Add a Beer
      tags:
      - beer
  /api/deletebeer:
    post:
      consumes:
      - application/json
      description: delete a beer
      parameters:
      - description: Delete Beer
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/string'
      summary: delete a Beer
      tags:
      - beer
  /api/getbeer/{id}:
    get:
      description: Get Beer for by beer_id
      parameters:
      - description: Beer Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200": {}
      summary: Get Beer
      tags:
      - beer
  /api/updatebeer:
    post:
      consumes:
      - application/json
      description: update a beer
      parameters:
      - description: Update Beer
        in: body
        name: beer
        required: true
        schema:
          $ref: '#/definitions/main.Beer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Beer'
      summary: update a Beer
      tags:
      - beer
swagger: "2.0"